+++ packages/language-css/src/index.ts
@@ -64,7 +64,7 @@ export type CssLanguagePluginOptions = {
  *   ],
  * });
  * @since v0.1.0
- * @version v0.1.24
+ * @version v0.1.26
  */
 export default class CssLanguagePlugin extends SimpleLanguagePlugin {
   /**